+++ pgversion.c
@@ -524,10 +524,10 @@ static char PgVersion_Type_Doc[] = "This is the type of PgVersion objects";
 
 PyTypeObject PgVersion_Type = {
 	PyObject_HEAD_INIT(NULL)
-	(int)NULL,							/* ob_size				*/
+	(Py_ssize_t)NULL,						/* ob_size				*/
 	MODULE_NAME ".PgVersion",			/* tp_name				*/
 	sizeof(PgVersion),					/* tp_basicsize			*/
-	(int)NULL,							/* tp_itemsize			*/
+	(Py_ssize_t)NULL,						/* tp_itemsize			*/
 	(destructor)PgVersion_dealloc,		/* tp_dealloc			*/
 	(printfunc)NULL,					/* tp_print				*/
 	(getattrfunc)PgVersion_getattr,		/* tp_getattr			*/
@@ -540,7 +540,7 @@ PyTypeObject PgVersion_Type = {
 	(hashfunc)NULL,						/* tp_hash				*/
 	(ternaryfunc)NULL,					/* tp_call				*/
 	(reprfunc)NULL,						/* tp_str				*/
-	NULL, NULL, NULL, (int)NULL,
+	NULL, NULL, NULL, (Py_ssize_t)NULL,
 	PgVersion_Type_Doc
 };
